+++ pkg/logging/db.go
@@ -9,7 +9,7 @@ import (
 
 	"gorm.io/gorm"
 
-	"github.com/jmpsec/osctrl/backend"
+	"github.com/jmpsec/osctrl/pkg/backend"
 	"github.com/jmpsec/osctrl/pkg/queries"
 	"github.com/jmpsec/osctrl/pkg/settings"
 	"github.com/jmpsec/osctrl/pkg/types"
